more test refactoring
