+++ packages/plugin-core/src/test/suite-integ/Extension.test.ts
@@ -30,8 +30,8 @@ import {
   describeMultiWS,
   runTestButSkipForWindows,
   setupBeforeAfter,
-  VSCodeTestUtils,
 } from "../testUtilsV3";
+import { VSCodeTestUtils } from "../utils";
 
 async function inactiveMessageTest(opts: {
   done: mocha.Done;
