+++ packages/plugin-core/src/test/testUtilsV3.ts
@@ -60,6 +60,7 @@ import {
   SetupWorkspaceOpts,
 } from "../commands/SetupWorkspace";
 import { VaultAddCommand } from "../commands/VaultAddCommand";
+import { ExtensionProvider } from "../ExtensionProvider";
 import { Logger } from "../logger";
 import { StateService } from "../services/stateService";
 import { WorkspaceConfig } from "../settings";
@@ -68,7 +69,6 @@ import { DendronExtension, getDWorkspace } from "../workspace";
 import { BlankInitializer } from "../workspace/blankInitializer";
 import { WorkspaceInitFactory } from "../workspace/WorkspaceInitFactory";
 import { _activate } from "../_extension";
-import { ExtensionProvider } from "../ExtensionProvider";
 import {
   cleanupVSCodeContextSubscriptions,
   setupCodeConfiguration,
@@ -77,7 +77,6 @@ import {
   stubWorkspaceFile,
   stubWorkspaceFolders,
 } from "./testUtilsv2";
-import * as vscode from "vscode";
 
 const TIMEOUT = 60 * 1000 * 5;
 
@@ -729,46 +728,4 @@ export async function waitInMilliseconds(milliseconds: number): Promise<void> {
       resolve();
     }, milliseconds);
   });
-}
-
-export class VSCodeTestUtils {
-  static mockUserConfigDir() {
-    const dir = tmpDir().name;
-    const getCodeUserConfigDurStub = sinon.stub(
-      VSCodeUtils,
-      "getCodeUserConfigDir"
-    );
-    getCodeUserConfigDurStub.callsFake(() => {
-      const wrappedMethod = getCodeUserConfigDurStub.wrappedMethod;
-      const originalOut = wrappedMethod();
-      return {
-        userConfigDir: [dir, originalOut.delimiter].join(""),
-        delimiter: originalOut.delimiter,
-        osName: originalOut.osName,
-      };
-    });
-    return getCodeUserConfigDurStub;
-  }
-
-  static stubWSFolders(wsRoot: string | undefined) {
-    if (wsRoot === undefined) {
-      const stub = sinon
-        .stub(vscode.workspace, "workspaceFolders")
-        .value(undefined);
-      DendronExtension.workspaceFolders = () => undefined;
-      return stub;
-    }
-    const wsFolders = [
-      {
-        name: "root",
-        index: 0,
-        uri: vscode.Uri.parse(wsRoot),
-      },
-    ];
-    const stub = sinon
-      .stub(vscode.workspace, "workspaceFolders")
-      .value(wsFolders);
-    DendronExtension.workspaceFolders = () => wsFolders;
-    return stub;
-  }
 }
