+++ packages/plugin-core/src/test/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./vscode";
+export * from "./workspace";
